Merge from Development to Master to prepare for Bali RC release

+++ ansible/script/keystone_multi.sh
@@ -63,44 +63,49 @@ DEV_STACK_LOCAL_CONF_DOCK
 chown stack:stack "$DEV_STACK_LOCAL_CONF"
 }
 
-opensds_conf() {
-cat >> "$OPENSDS_CONFIG_DIR/opensds.conf" << OPENSDS_GLOBAL_CONFIG_DOC
+multi_conf() {
+cat >> "$MULTI_CONFIG_DIR/multi.conf" << MULTI_GLOBAL_CONFIG_DOC
 
 
 [keystone_authtoken]
 memcached_servers = $HOST_IP:11211
-signing_dir = /var/cache/opensds
+signing_dir = /var/cache/multi
 cafile = /opt/stack/data/ca-bundle.pem
 auth_uri = http://$HOST_IP/identity
 project_domain_name = Default
 project_name = service
 user_domain_name = Default
 password = $STACK_PASSWORD
-username = $OPENSDS_SERVER_NAME
+username = $MULTI_SERVER_NAME
 auth_url = http://$HOST_IP/identity
 auth_type = password
 
-OPENSDS_GLOBAL_CONFIG_DOC
+MULTI_GLOBAL_CONFIG_DOC
 
-cp "$ANSIBLE_CONF_DIR/policy.json" "$OPENSDS_CONFIG_DIR"
 }
 
 create_user_and_endpoint(){
     . "$DEV_STACK_DIR/openrc" admin admin
-    openstack user create --domain default --password "$STACK_PASSWORD" "$OPENSDS_SERVER_NAME"
-    openstack role add --project service --user "$OPENSDS_SERVER_NAME" admin
+    if openstack user show $MULTI_SERVER_NAME &>/dev/null; then
+        return 
+    fi
+    openstack user create --domain default --password "$STACK_PASSWORD" "$MULTI_SERVER_NAME"
+    openstack role add --project service --user "$MULTI_SERVER_NAME" admin
     openstack group create service
-    openstack group add user service "$OPENSDS_SERVER_NAME"
+    openstack group add user service "$MULTI_SERVER_NAME"
     openstack role add service --project service --group service
     openstack group add user admins admin
-    openstack service create --name "opensds$OPENSDS_VERSION" --description "OpenSDS Block Storage" "opensds$OPENSDS_VERSION"
-    openstack endpoint create --region RegionOne "opensds$OPENSDS_VERSION" public "http://$HOST_IP:50040/$OPENSDS_VERSION/%(tenant_id)s"
-    openstack endpoint create --region RegionOne "opensds$OPENSDS_VERSION" internal "http://$HOST_IP:50040/$OPENSDS_VERSION/%(tenant_id)s"
-    openstack endpoint create --region RegionOne "opensds$OPENSDS_VERSION" admin "http://$HOST_IP:50040/$OPENSDS_VERSION/%(tenant_id)s"
+    openstack service create --name "multicloud$MULTI_VERSION" --description "Multi-cloud Block Storage" "multicloud$MULTI_VERSION"
+    openstack endpoint create --region RegionOne "multicloud$MULTI_VERSION" public "http://$HOST_IP:8089/$MULTI_VERSION/%(tenant_id)s"
+    openstack endpoint create --region RegionOne "multicloud$MULTI_VERSION" internal "http://$HOST_IP:8089/$MULTI_VERSION/%(tenant_id)s"
+    openstack endpoint create --region RegionOne "multicloud$MULTI_VERSION" admin "http://$HOST_IP:8089/$MULTI_VERSION/%(tenant_id)s"
 }
 
 delete_redundancy_data() {
     . "$DEV_STACK_DIR/openrc" admin admin
+    if ! openstack user show demo &>/dev/null; then
+        return
+    fi
     openstack project delete demo
     openstack project delete alt_demo
     openstack project delete invisible_to_admin
@@ -118,15 +123,14 @@ download_code(){
 install(){
     create_user
     download_code
-    opensds_conf
+    multi_conf
 
     # If keystone is ready to start, there is no need continue next step.
-    if wait_for_url "http://$HOST_IP/identity" "keystone" 0.25 4; then
-        return
+    if ! wait_for_url "http://$HOST_IP/identity" "keystone" 0.25 4; then
+        devstack_local_conf
+        cd "${DEV_STACK_DIR}"
+        su "$STACK_USER_NAME" -c "${DEV_STACK_DIR}/stack.sh" >/dev/null
     fi
-    devstack_local_conf
-    cd "${DEV_STACK_DIR}"
-    su "$STACK_USER_NAME" -c "${DEV_STACK_DIR}/stack.sh" >/dev/null
     create_user_and_endpoint
     delete_redundancy_data
 }
@@ -147,8 +151,8 @@ uninstall_purge(){
 # ***************************
 TOP_DIR=$(cd $(dirname "$0") && pwd)
 
-# OpenSDS configuration directory
-OPENSDS_CONFIG_DIR=${OPENSDS_CONFIG_DIR:-/etc/opensds}
+# Multi-cloud configuration directory
+MULTI_CONFIG_DIR=${MULTI_CONFIG_DIR:-/etc/multi}
 
 source "$TOP_DIR/util.sh"
 source "$TOP_DIR/sdsrc"
